refactor(web): dissolve domains/nudges/ into shared top-level directories (LUM-1919)
Dissolves the nudges domain entirely — every file consumed by 2+ domains
now lives in a top-level shared directory:

- Platform detection (isIOSBrowser, isMacOSBrowser, useIsIOSWeb,
  useIsMacOSWeb) → utils/platform-detection.ts
- Nudge Zustand store → stores/nudge-store.ts
- GitHub/Discord/iOS/macOS nudge hooks + prefs → hooks/use-*-nudge.ts
- localStorage helpers → utils/nudge-prefs.ts
- Banner UI components → components/nudges/

Constants merged into their respective hook files to reduce file count;
legacy-cleanup key strings inlined in the store to avoid circular deps.

7 cross-domain violations eliminated (95 → 88).

Closes LUM-1919
