+++ apps/web/src/components/nudges/macos-app-banner.tsx
@@ -2,7 +2,7 @@
 import { Download } from "lucide-react";
 
 import { AppleLogo } from "@/components/icons/apple-logo.js";
-import { NudgeChatBanner } from "@/domains/nudges/components/nudge-chat-banner.js";
+import { NudgeChatBanner } from "@/components/nudges/nudge-chat-banner.js";
 
 interface MacOSAppBannerProps {
   onDownload: () => void;
