--- apps/web/src/domains/nudges/ios-app-constants.ts
@@ -1,15 +1,0 @@
-/** localStorage key: user tapped "Download" on any iOS nudge surface. */
-export const KEY_IOS_APP_DOWNLOADED = "app.iosNudge.downloaded";
-
-/** localStorage key: user dismissed the in-chat floating banner. */
-export const KEY_IOS_APP_BANNER_DISMISSED = "app.iosNudge.bannerDismissed";
-
-/** localStorage key: cumulative completed assistant turns observed on web. */
-export const KEY_IOS_APP_ASSISTANT_TURNS_SEEN =
-  "app.iosNudge.assistantTurnsSeen";
-
-export const IOS_APP_BANNER_MIN_TURNS = 5;
-
-/** App Store listing for Vellum Assistant (id6759934423). */
-export const IOS_APP_STORE_URL =
-  "https://apps.apple.com/us/app/vellum-assistant/id6759934423";
