+++ apps/web/src/domains/settings/components/ios-app-card.tsx
@@ -1,11 +1,11 @@
 import { Bell, Fingerprint, Smartphone, Vibrate } from "lucide-react";
 
 import { NudgeSettingsCard } from "@/domains/settings/components/nudge-settings-card.js";
-import { useIsIOSWeb } from "@/domains/nudges/ios-app-platform.js";
+import { useIsIOSWeb } from "@/utils/platform-detection.js";
 import {
   openIOSAppStore,
   writeIOSAppDownloaded,
-} from "@/domains/nudges/ios-app-prefs.js";
+} from "@/hooks/use-ios-app-nudge.js";
 
 export function IOSAppCard() {
   const isIOSWeb = useIsIOSWeb();
