+++ apps/web/src/hooks/use-macos-app-nudge.ts
@@ -1,3 +1,11 @@
+/**
+ * macOS app-download nudge module.
+ *
+ * Manages whether the user has downloaded the macOS app, banner
+ * dismissal state, and assistant turn counting for the minimum-turn
+ * threshold before the nudge banner surfaces.
+ */
+
 import { useCallback, useEffect, useState } from "react";
 
 import {
@@ -5,14 +13,29 @@ import {
   writeBooleanPref,
   readNumberPref,
   writeNumberPref,
-} from "@/domains/nudges/nudge-prefs.js";
+} from "@/utils/nudge-prefs.js";
 
-import {
-  KEY_MAC_APP_DOWNLOADED,
-  KEY_MAC_APP_BANNER_DISMISSED,
-  KEY_MAC_APP_ASSISTANT_TURNS_SEEN,
-  MACOS_DOWNLOAD_URL,
-} from "@/domains/nudges/mac-app-constants.js";
+// ---------------------------------------------------------------------------
+// Constants
+// ---------------------------------------------------------------------------
+
+/** localStorage key: user clicked "Download" on any nudge surface. */
+export const KEY_MAC_APP_DOWNLOADED = "app.macOsNudge.downloaded";
+
+/** localStorage key: user dismissed the in-chat floating banner. */
+export const KEY_MAC_APP_BANNER_DISMISSED = "app.macOsNudge.bannerDismissed";
+
+/** localStorage key: cumulative completed assistant turns observed on web. */
+export const KEY_MAC_APP_ASSISTANT_TURNS_SEEN =
+  "app.macOsNudge.assistantTurnsSeen";
+
+export const MAC_APP_BANNER_MIN_TURNS = 5;
+
+/**
+ * macOS app download URL. Replace with the canonical CDN or marketing
+ * page URL before shipping.
+ */
+export const MACOS_DOWNLOAD_URL = "https://vellum.ai/download";
 
 // ---------------------------------------------------------------------------
 // Public readers / writers
@@ -45,7 +68,7 @@ export function incrementMacOsAssistantTurnsSeen(delta = 1): void {
 }
 
 // ---------------------------------------------------------------------------
-// Hooks
+// Hook
 // ---------------------------------------------------------------------------
 
 export function useMacOsNudgeState(): {
@@ -86,16 +109,3 @@ export function useMacOsNudgeState(): {
 export function openMacOsDownload(): void {
   window.open(MACOS_DOWNLOAD_URL, "_blank", "noopener,noreferrer");
 }
-
-// ---------------------------------------------------------------------------
-// Internals exported for tests only. Not part of the public API.
-// ---------------------------------------------------------------------------
-
-export const __testing = {
-  readBooleanPref,
-  writeBooleanPref,
-  readNumberPref,
-  writeNumberPref,
-  readMacOsAppBannerDismissed,
-  writeMacOsAppBannerDismissed,
-};
