+++ apps/web/src/stores/nudge-store.ts
@@ -2,9 +2,8 @@
  * Zustand store for GitHub + Discord nudge prefs.
  *
  * Owns whether each nudge has been actioned (starred, joined) or
- * dismissed (banner) and when. `github-prefs.ts` and
- * `discord-prefs.ts` expose thin selector hooks (`useGitHubNudgeState`,
- * `useDiscordNudgeState`) backed by this store.
+ * dismissed (banner) and when. `use-github-nudge.ts` and
+ * `use-discord-nudge.ts` expose thin selector hooks backed by this store.
  *
  * **Storage model:**
  *
@@ -23,17 +22,6 @@ import { create } from "zustand";
 import { persist, createJSONStorage } from "zustand/middleware";
 
 import { createSelectors } from "@/utils/create-selectors.js";
-
-import {
-  KEY_GITHUB_NUDGE_STARRED,
-  KEY_GITHUB_NUDGE_BANNER_DISMISSED,
-  KEY_GITHUB_NUDGE_BANNER_DISMISSED_AT,
-} from "@/domains/nudges/github-constants.js";
-import {
-  KEY_DISCORD_NUDGE_JOINED,
-  KEY_DISCORD_NUDGE_BANNER_DISMISSED,
-  KEY_DISCORD_NUDGE_FIRST_SEEN_AT,
-} from "@/domains/nudges/discord-constants.js";
 
 // ---------------------------------------------------------------------------
 // State + Actions
@@ -102,7 +90,6 @@ const useNudgeStoreBase = create<NudgeStore>()(
     {
       name: NUDGE_STORE_KEY,
       storage: createJSONStorage(() => localStorage),
-      // Only persist state, not action functions.
       partialize: (state) => ({
         githubStarred: state.githubStarred,
         githubBannerDismissed: state.githubBannerDismissed,
@@ -121,9 +108,6 @@ export const useNudgeStore = createSelectors(useNudgeStoreBase);
 // Cross-tab sync
 // ---------------------------------------------------------------------------
 
-// `localStorage.setItem` fires a native `storage` event in *other* tabs.
-// Persist middleware doesn't subscribe to it on its own, so wire a listener
-// that rehydrates this store whenever `vellum:nudge-prefs` changes elsewhere.
 if (typeof window !== "undefined") {
   window.addEventListener("storage", (event) => {
     if (event.key === NUDGE_STORE_KEY) {
@@ -136,28 +120,20 @@ if (typeof window !== "undefined") {
 // One-shot legacy cleanup
 // ---------------------------------------------------------------------------
 
-// One-time cleanup of legacy per-key localStorage entries from an
-// older nudge-state shape. The persist middleware now owns the state
-// under `vellum:nudge-prefs`; the old keys are orphaned bytes on
-// every user's device and get removed on first load.
 const LEGACY_CLEANUP_FLAG = "app.nudgeLegacy.cleaned";
 
 const LEGACY_KEYS = [
-  KEY_GITHUB_NUDGE_STARRED,
-  KEY_GITHUB_NUDGE_BANNER_DISMISSED,
-  KEY_GITHUB_NUDGE_BANNER_DISMISSED_AT,
-  KEY_DISCORD_NUDGE_JOINED,
-  KEY_DISCORD_NUDGE_BANNER_DISMISSED,
-  KEY_DISCORD_NUDGE_FIRST_SEEN_AT,
+  "app.githubNudge.starred",
+  "app.githubNudge.bannerDismissed",
+  "app.githubNudge.bannerDismissedAt",
+  "app.discordNudge.joined",
+  "app.discordNudge.bannerDismissed",
+  "app.discordNudge.firstSeenAt",
 ];
 
 if (typeof window !== "undefined") {
   try {
     if (localStorage.getItem(LEGACY_CLEANUP_FLAG) !== "true") {
-      // Calling localStorage.removeItem directly (rather than going through
-      // `removeLocalSetting` in domains/settings/) keeps this cleanup
-      // self-contained — nothing listens for the `vellum:pref-changed`
-      // event on these specific legacy keys.
       for (const key of LEGACY_KEYS) {
         localStorage.removeItem(key);
       }
